+++ config/webpack/plugin/TerraDevSiteEntrypoints.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+/**
+ * Default entries for terra-dev-site
+ */
+module.exports = {
+  rewriteHistory: path.resolve(path.join(__dirname, '..', '..', '..', 'lib', 'rewriteHistory')),
+  redirect: path.resolve(path.join(__dirname, '..', '..', '..', 'lib', 'redirect')),
+};