+++ config/webpack/webpack.config.js
@@ -4,9 +4,7 @@ const terraDevSiteConfig = require('./terra-dev-site.webpack.config');
 
 
 const mergedConfig = (env, argv) => (
-  merge.strategy({
-    'resolve.modules': 'prepend',
-  })(defaultWebpackConfig(env, argv), terraDevSiteConfig(env, argv))
+  merge(defaultWebpackConfig(env, argv), terraDevSiteConfig(env, argv))
 );
 
 module.exports = mergedConfig;