+++ scripts/generate-app-config/loadSiteConfig.js
@@ -1,16 +1,15 @@
 const path = require('path');
 const fs = require('fs');
-const defaultConfig = require('../../config/site/site.config.js');
 
 /**
-* Detemine if the path resolves to a file.
+* Determine if the path resolves to a file.
 */
 const isFile = filePath => (fs.existsSync(filePath) && !fs.lstatSync(filePath).isDirectory());
 
 /**
 * Overlay the default site config on the custom site config.
 */
-const applyDefaults = (filePath) => {
+const applyDefaults = (filePath, defaultConfig) => {
   // eslint-disable-next-line global-require, import/no-dynamic-require
   const config = require(filePath);
 
@@ -22,9 +21,9 @@ const applyDefaults = (filePath) => {
 /**
 * If the file path is valid, merge the custom site config with the defaults.
 */
-const resolve = (filePath) => {
+const resolve = (filePath, defaultConfig) => {
   if (isFile(filePath)) {
-    return applyDefaults(filePath);
+    return applyDefaults(filePath, defaultConfig);
   }
 
   return undefined;
@@ -33,16 +32,14 @@ const resolve = (filePath) => {
 
 /**
  * Returns the site configuration. It will attempt to load the configuration from the provided configPath first or
- * the defualt config path and then merge with the default config or it will just return the default site config.
+ * the default config path and then merge with the default config or it will just return the default site config.
  */
-const loadSiteConfig = (configPath) => {
-  // Merge the provided site config with the defaults
-  if (configPath) {
-    return applyDefaults(path.resolve(configPath));
-  }
+const loadSiteConfig = (configName = 'site.config.js', defaultPath = '../../config/site/site.config.js') => {
+  // eslint-disable-next-line global-require, import/no-dynamic-require
+  const defaultConfig = require(defaultPath);
 
-  // Try to find the site config at the defualt path and then merge it with the defaults.
-  const localConfig = resolve(path.resolve(process.cwd(), 'dev-site-config', 'site.config.js'));
+  // Try to find the site config at the default path and then merge it with the defaults.
+  const localConfig = resolve(path.resolve(process.cwd(), 'dev-site-config', configName), defaultConfig);
   if (localConfig) {
     return localConfig;
   }